+++ src/ByteSequenceTest.php
@@ -4,17 +4,16 @@ declare(strict_types=1);
 
 namespace Bakame\Http\StructuredFields;
 
-/**
- * @coversDefaultClass \Bakame\Http\StructuredFields\ByteSequence
- */
-final class ByteSequenceTest extends StructuredFieldTest
+use PHPUnit\Framework\Attributes\Test;
+
+final class ByteSequenceTest extends StructuredFieldTestCase
 {
     /** @var array<string> */
-    protected array $paths = [
+    protected static array $paths = [
         __DIR__.'/../vendor/httpwg/structured-field-tests/binary.json',
     ];
 
-    /** @test */
+    #[Test]
     public function it_will_fail_on_invalid_decoded_string_with_inner_space(): void
     {
         $this->expectException(SyntaxError::class);
@@ -22,7 +21,7 @@ final class ByteSequenceTest extends StructuredFieldTest
         ByteSequence::fromEncoded('a a');
     }
 
-    /** @test */
+    #[Test]
     public function it_will_fail_on_invalid_decoded_string(): void
     {
         $this->expectException(SyntaxError::class);
@@ -30,7 +29,7 @@ final class ByteSequenceTest extends StructuredFieldTest
         ByteSequence::fromEncoded('aaaaa');
     }
 
-    /** @test */
+    #[Test]
     public function it_can_decode_base64_field(): void
     {
         $source = 'cHJldGVuZCB0aGlzIGlzIGJpbmFyeSBjb250ZW50Lg==';
@@ -40,7 +39,7 @@ final class ByteSequenceTest extends StructuredFieldTest
         self::assertSame($source, $item->encoded());
     }
 
-    /** @test */
+    #[Test]
     public function it_can_encode_raw_field(): void
     {
         $decoded = 'pretend this is binary content.';
@@ -51,7 +50,7 @@ final class ByteSequenceTest extends StructuredFieldTest
         self::assertSame($source, $item->encoded());
     }
 
-    /** @test */
+    #[Test]
     public function it_can_compare_instances(): void
     {
         $decoded = 'pretend this is binary content.';