+++ src/ParserTest.php
@@ -5,17 +5,18 @@ declare(strict_types=1);
 namespace Bakame\Http\StructuredFields;
 
 use DateTimeImmutable;
+use PHPUnit\Framework\Attributes\Test;
 
-final class ParserTest extends StructuredFieldTest
+final class ParserTest extends StructuredFieldTestCase
 {
     /** @var array<string> */
-    protected array $paths = [
+    protected static array $paths = [
         __DIR__.'/../vendor/httpwg/structured-field-tests/examples.json',
         __DIR__.'/../vendor/httpwg/structured-field-tests/key-generated.json',
         __DIR__.'/../vendor/httpwg/structured-field-tests/large-generated.json',
     ];
 
-    /** @test */
+    #[Test]
     public function it_will_fail_with_wrong_boolean(): void
     {
         $this->expectException(SyntaxError::class);
@@ -23,7 +24,7 @@ final class ParserTest extends StructuredFieldTest
         Parser::parseDictionary('a=1, b=?3;foo=9, c=3');
     }
 
-    /** @test */
+    #[Test]
     public function it_will_fail_with_wrong_number(): void
     {
         $this->expectException(SyntaxError::class);
@@ -31,7 +32,7 @@ final class ParserTest extends StructuredFieldTest
         Parser::parseDictionary('key=-1ab');
     }
 
-    /** @test */
+    #[Test]
     public function it_will_fail_with_wrong_sequence(): void
     {
         $this->expectException(SyntaxError::class);
@@ -39,7 +40,7 @@ final class ParserTest extends StructuredFieldTest
         Parser::parseDictionary('a=:toto89é:');
     }
 
-    /** @test */
+    #[Test]
     public function it_parse_a_date_item(): void
     {
         $field = Parser::parseDictionary('a=@12345678;key=1');
@@ -48,7 +49,7 @@ final class ParserTest extends StructuredFieldTest
         self::assertInstanceOf(DateTimeImmutable::class, $field['a']->value());
     }
 
-    /** @test */
+    #[Test]
     public function it_will_fail_with_wrong_date_format(): void
     {
         $this->expectException(SyntaxError::class);
@@ -56,7 +57,7 @@ final class ParserTest extends StructuredFieldTest
         Parser::parseDictionary('a=@12345.678');
     }
 
-    /** @test */
+    #[Test]
     public function it_will_fail_with_out_of_range_date_format(): void
     {
         $this->expectException(SyntaxError::class);
@@ -64,7 +65,7 @@ final class ParserTest extends StructuredFieldTest
         Parser::parseDictionary('a=@'. 1_000_000_000_000_000);
     }
 
-    /** @test */
+    #[Test]
     public function it_will_fail_with_wrong_string(): void
     {
         $this->expectException(SyntaxError::class);
@@ -72,7 +73,7 @@ final class ParserTest extends StructuredFieldTest
         Parser::parseDictionary('a="foo \O bar"');
     }
 
-    /** @test */
+    #[Test]
     public function it_will_fail_with_wrong_string_utf8(): void
     {
         $this->expectException(SyntaxError::class);
@@ -80,7 +81,7 @@ final class ParserTest extends StructuredFieldTest
         Parser::parseDictionary('a="foébar"');
     }
 
-    /** @test */
+    #[Test]
     public function it_fails_to_parse_invalid_string_1(): void
     {
         $this->expectException(SyntaxError::class);
@@ -88,7 +89,7 @@ final class ParserTest extends StructuredFieldTest
         Parser::parseList('(foo;number="hello\")');
     }
 
-    /** @test */
+    #[Test]
     public function it_fails_to_parse_invalid_string_2(): void
     {
         $this->expectException(SyntaxError::class);
@@ -96,7 +97,7 @@ final class ParserTest extends StructuredFieldTest
         Parser::parseDictionary('number="hell\o"');
     }
 
-    /** @test */
+    #[Test]
     public function it_fails_to_parse_an_invalid_http_field(): void
     {
         $this->expectException(SyntaxError::class);
@@ -104,7 +105,7 @@ final class ParserTest extends StructuredFieldTest
         Parser::parseInnerList('("hello)world" 42 42.0;john=doe);foo="bar(" toto');
     }
 
-    /** @test */
+    #[Test]
     public function it_fails_to_parse_an_invalid_http_field_2(): void
     {
         $this->expectException(SyntaxError::class);